@@ -1,12 +1,12 @@
 [package]
 name = "xtask"
-version = "0.15.4"
-edition = "2021"
+version = "0.1.1"
+edition = "2024"
 
 [dependencies]
-clap = { version = "4.5.51", features = ["derive"] }
-clap_complete = "4.5.60"
-clap_complete_fig = "4.5.2"
-clap_complete_nushell = "4.5.10"
-clap_mangen = "0.2.31"
+clap = { workspace = true, features = ["derive", "unstable-markdown"] }
+clap_complete = { workspace = true }
+clap_complete_fig = { workspace = true }
+clap_complete_nushell = { workspace = true }
+clap_mangen = { workspace = true }
 skim = { path = "../skim", features = [] }
