@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use std::sync::Mutex;
+use std::sync::{LazyLock, Mutex};
 use std::sync::atomic::{AtomicU32, Ordering};
 
 // Consider that you invoke a command with different arguments several times
@@ -7,15 +7,13 @@ use std::sync::atomic::{AtomicU32, Ordering};
 // => Well, we'll give each invocation a number, i.e. RUN_NUM
 // What if you invoke the same command and same arguments twice?
 // => We use NUM_MAP to specify the same run number.
-lazy_static! {
-    static ref RUN_NUM: AtomicU32 = AtomicU32::new(0);
-    static ref SEQ: AtomicU32 = AtomicU32::new(1);
-    static ref NUM_MAP: Mutex<HashMap<String, u32>> = {
-        let mut m = HashMap::new();
-        m.insert("".to_string(), 0);
-        Mutex::new(m)
-    };
-}
+static RUN_NUM: LazyLock<AtomicU32> = LazyLock::new(|| AtomicU32::new(0));
+static SEQ: LazyLock<AtomicU32> = LazyLock::new(|| AtomicU32::new(1));
+static NUM_MAP: LazyLock<Mutex<HashMap<String, u32>>> = LazyLock::new(|| {
+    let mut m = HashMap::new();
+    m.insert("".to_string(), 0);
+    Mutex::new(m)
+});
 
 pub fn current_run_num() -> u32 {
     RUN_NUM.load(Ordering::SeqCst)
