@@ -6,7 +6,8 @@ extern crate skim;
 extern crate time;
 
 use crate::Event;
-use clap::{Error, Parser};
+use clap::{CommandFactory, Error, Parser};
+use clap_complete::generate;
 use color_eyre::Result;
 use color_eyre::eyre::eyre;
 use derive_builder::Builder;
@@ -90,6 +91,13 @@ fn main() -> Result<()> {
 fn sk_main() -> Result<i32> {
     let mut opts = parse_args()?.build();
 
+    // Handle shell completion generation if requested
+    if let Some(shell) = opts.shell {
+        // Generate completion script directly to stdout
+        generate(shell, &mut SkimOptions::command(), "sk", &mut io::stdout());
+        return Ok(0);
+    }
+
     let reader_opts = SkimItemReaderOption::default()
         .ansi(opts.ansi)
         .delimiter(opts.delimiter.clone())
@@ -118,7 +126,7 @@ fn sk_main() -> Result<i32> {
     //------------------------------------------------------------------------------
     // output
 
-    let Some(result) = (if opts.tmux.is_some() {
+    let Some(result) = (if opts.tmux.is_some() && env::var("TMUX").is_ok() {
         crate::tmux::run_with(&opts)
     } else {
         // read from pipe or command
@@ -250,12 +258,13 @@ pub fn filter(bin_option: &BinOptions, options: &SkimOptions, source: Option<Ski
     });
 
     let mut num_matched = 0;
+    let mut stdout_lock = std::io::stdout().lock();
     stream_of_item
         .into_iter()
         .filter_map(|item| engine.match_item(item.clone()).map(|result| (item, result)))
         .for_each(|(item, _match_result)| {
             num_matched += 1;
-            print!("{}{}", item.output(), bin_option.output_ending);
+            let _ = write!(stdout_lock, "{}{}", item.output(), bin_option.output_ending);
         });
 
     i32::from(num_matched == 0)
