@@ -15,12 +15,12 @@ fn setup_tmux_mock(tmux: &TmuxController) -> Result<String> {
     let dir = &tmux.tempdir;
     let path = dir.path().join("tmux");
     let mock_bin = Path::new(&path);
-    let mut writer = File::create_new(mock_bin)?;
+    let mut writer = File::create(mock_bin)?;
     let outfile = dir.path().join("tmux-mock-cmd");
     writer.write_fmt(format_args!(
         "#!/bin/sh
 
-echo $@ > {}
+echo \"$@\" > {}
 ",
         outfile.to_str().unwrap()
     ))?;
@@ -63,6 +63,64 @@ fn tmux_stdin() -> Result<()> {
     let cmd = get_tmux_cmd(&outfile)?;
     println!("{}", cmd);
     assert!(cmd.contains("<"));
+
+    Ok(())
+}
+
+#[test]
+fn tmux_quote_bash() -> Result<()> {
+    let tmux = TmuxController::new()?;
+    let outfile = setup_tmux_mock(&tmux)?;
+    tmux.send_keys(&[Str("export SHELL=/bin/bash"), Enter])?;
+    tmux.start_sk(None, &["--tmux", "--bind 'ctrl-a:reload(ls /foo*)'"])?;
+    tmux.until(|_| Path::new(&outfile).exists())?;
+    let cmd = get_tmux_cmd(&outfile)?;
+    assert!(cmd.starts_with("display-popup"));
+    assert!(cmd.contains("-E"));
+    assert!(cmd.contains("--bind $'ctrl-a:reload(ls /foo*)'"));
+
+    Ok(())
+}
+#[test]
+fn tmux_quote_zsh() -> Result<()> {
+    let tmux = TmuxController::new()?;
+    let outfile = setup_tmux_mock(&tmux)?;
+    tmux.send_keys(&[Str("export SHELL=/bin/zsh"), Enter])?;
+    tmux.start_sk(None, &["--tmux", "--bind 'ctrl-a:reload(ls /foo*)'"])?;
+    tmux.until(|_| Path::new(&outfile).exists())?;
+    let cmd = get_tmux_cmd(&outfile)?;
+    println!("{cmd}");
+    assert!(cmd.starts_with("display-popup"));
+    assert!(cmd.contains("-E"));
+    assert!(cmd.contains("sk --bind $'ctrl-a:reload(ls /foo*)' >"));
+
+    Ok(())
+}
+#[test]
+fn tmux_quote_sh() -> Result<()> {
+    let tmux = TmuxController::new()?;
+    let outfile = setup_tmux_mock(&tmux)?;
+    tmux.send_keys(&[Str("export SHELL=/bin/sh"), Enter])?;
+    tmux.start_sk(None, &["--tmux", "--bind 'ctrl-a:reload(ls /foo*)'"])?;
+    tmux.until(|_| Path::new(&outfile).exists())?;
+    let cmd = get_tmux_cmd(&outfile)?;
+    assert!(cmd.starts_with("display-popup"));
+    assert!(cmd.contains("-E"));
+    assert!(cmd.contains("--bind ctrl-a':reload(ls /foo*)'"));
+
+    Ok(())
+}
+#[test]
+fn tmux_quote_fish() -> Result<()> {
+    let tmux = TmuxController::new()?;
+    let outfile = setup_tmux_mock(&tmux)?;
+    tmux.send_keys(&[Str("export SHELL=/bin/sh"), Enter])?;
+    tmux.start_sk(None, &["--tmux", "--bind 'ctrl-a:reload(ls /foo*)'"])?;
+    tmux.until(|_| Path::new(&outfile).exists())?;
+    let cmd = get_tmux_cmd(&outfile)?;
+    assert!(cmd.starts_with("display-popup"));
+    assert!(cmd.contains("-E"));
+    assert!(cmd.contains("--bind ctrl-a':reload(ls /foo*)'"));
 
     Ok(())
 }
